+++ lib/searchkick/record_indexer.rb
@@ -59,7 +59,10 @@ module Searchkick
       routing = items.to_h { |r| [r[:id], r[:routing]] }
       record_ids = routing.keys
 
-      records = Searchkick.load_records(klass, record_ids).select(&:should_index?)
+      relation = Searchkick.load_records(klass, record_ids)
+      # call search_import even for single records for nested associations
+      relation = relation.search_import if relation.respond_to?(:search_import)
+      records = relation.select(&:should_index?)
 
       # determine which records to delete
       delete_ids = record_ids - records.map { |r| r.id.to_s }